Factor out polygon validity code into its own file.
